relocated queries, changed names, fixed imports
