+++ src/MainContainer/ContentContainer/Company/SelectCompany.js
@@ -3,7 +3,7 @@ import Language from '../../../utils/language'
 import PropTypes from 'prop-types'
 import React, { useState, useContext } from 'react'
 import { useMutation } from 'react-apollo-hooks'
-import { SET_ACTIVE_COMPANY } from '../../../utils/query'
+import { SET_ACTIVE_COMPANY } from '../../../utils/Query/CompanyQuery'
 import {
   withStyles,
   InputLabel,
