+++ src/MainContainer/ContentContainer/Invoice/Logic.js
@@ -2,7 +2,7 @@ import { useState, useContext } from 'react'
 import Context from '../../../Context/Context'
 import { GET_PRODUCTS } from '../../../utils/Query/ProductQuery'
 import { GET_CUSTOMERS } from '../../../utils/Query/CustomersQuery'
-import { GET_SUBSCRIP_ACCOUNTS } from '../../../utils/query'
+import { GET_SUBSCRIP_ACCOUNTS } from '../../../utils/Query/AccountQuery'
 import { POST_INVOICE } from '../../../utils/Query/InvoiceQuery'
 import { POST_ORDER } from '../../../utils/Query/OrderQuery'
 import { useSubscription, useMutation } from 'react-apollo-hooks'
